+++ src/modules/0.0.21/iasql_functions/index.ts
@@ -1,21 +1,19 @@
 /* THIS MODULE IS A SPECIAL SNOWFLAKE. DON'T LOOK AT IT FOR HOW TO WRITE A REAL MODULE */
-import { Context, ModuleBase } from '../../interfaces';
+import { ModuleBase } from '../../interfaces';
 import { IasqlOperationType } from './entity';
+import { CustomCallRpc } from './rpcs';
 
-class IasqlFunctions extends ModuleBase {
+export class IasqlFunctions extends ModuleBase {
+  customCall: CustomCallRpc;
   constructor() {
     super();
     this.sql = {
       afterInstallSqlPath: 'sql/create_fns.sql',
       beforeUninstallSqlPath: 'sql/drop_fns.sql',
     };
+    this.customCall = new CustomCallRpc(this);
     super.init();
   }
   iasqlOperationType = IasqlOperationType;
-  // TODO: here for testing purpose. To be delete it
-  // TODO: this also will be refactored using an rpc subtype in ModuleBase
-  async customCall(ctx: Context, arg1: string, arg2: string) {
-    return JSON.stringify([{ result: 'I have been called!', arg1, arg2, memo: ctx.memo }]);
-  }
 }
 export const iasqlFunctions = new IasqlFunctions();
